Added comments and JavaDocs to some Classes
Added comments and JavaDocs to the Food, GameUtil, MusicPlayer, Play and Snake classes.

Each comment/JavaDoc defines what each method in each class does.

I added these comments/JavaDocs to give a better understanding of the code. These are just basic evaluations of the methods and may potentially be added to in the future with more detail.