Added JavaFx class
Added a JavaFx class for the start screen and set up 2 buttons. A button that will start the game, and a button that will take you to a choice of colours. The colour chosen is the colour for the text that dsiplays the score.

More will be added to the start screen, this is just a basis and gettng it setup.