feat/refactor: Added fxml files and Controllers instead of using JavaFx file
I added fxml files and controllers for these fxml files instead of the generic and clunky JavaFx file I had. Now the code is a lot lighter and easier to follow.

I put all these files into a Package called Application

For now, I am keeping the JavaFx file just incase but it will be deleted in the future.