feat: Added functionality to Start button in Start Screen
When the start button is pressed, the current screen is no longer visible and the Play Screen becomes visible, ready for the user to play the game.

The colour of the score is also the same as requested in the start screen, this is the same with the background that is chosen.