feat: Add functionality to buttons on the End Screen
The 'Play Again' button now restarts the game, taking the user back to the start screen. The 'Quit' button now closes the application.

There is a temporary button in the Play Screen until I implement a moving snake in JavaFx so I am able to check the End Screen works correctly