chore: Created some new function for duplicate code
Any code that was often duplicated, I created a function for.

For example, there was a few times where the code created bounds and then checked if the snake head intersected with these bounds. I put this into a function CheckBounds where a Node is sent and the function checks if the snakehead intersects with the given Node.