fix/refactor: Changed the Wall, Bomb and Food classes
Walls, Bombs and Food could still be out of bounds sometimes. I have now correctly fixed this.

I removed the x and y parameters from the constructor for all 3 classes because it was not necessary to send 2 random integers. The class can just get the x and y boundaries then create the random integers themselves.